Improved notebook
Fixed small things such as the for loop filling up missing pictures (which ran from 0 to 19998) (now runs from 1 to 20000). Also made the Images class independent of the variable labels, it can now be
created using only the csv file. This allows to split up the notebook in a preparation part and a seperate part to train the model. Added a test set to the train and validation set.