New model and test data
Added code to process the test dataset and predict its labels. Also the
neural net is new, with more convolutions etc. To save time it was just
trained for one epoch and probably did not learn much, that's why the
predictions  for test data are all 0. This was more to check if the test
prediction routines work as planned, which they do.